Finished most of unit 5

unit 5 complete

bug fix for lab 1

?

??

+++ src/unit4/Lab4_Act7.py
@@ -11,11 +11,11 @@ import math
 def main():
     # Reads in input
     print('Please input A.')
-    a = float(input())
+    a = int(input())
     print('Input B')
-    b = float(input())
+    b = int(input())
     print('Input C')
-    c = float(input())
+    c = int(input())
 
     # Checks if A is zero
     if a == 0:
